refactor(web): streamline SecurityPanelView component by removing unused imports and consolidating logic into primitives
