+++ apps/web/src/features/spaces/components/SecurityHub/components/SecurityPanelView/SecurityPanelView.tsx
@@ -1,21 +1,24 @@
 import { type ReactElement, type ReactNode, useMemo, useState } from 'react'
-import { Box, Button, CircularProgress, Collapse, Divider, Paper, Skeleton, Stack, Typography } from '@mui/material'
+import { Box, CircularProgress, Collapse, Divider, Paper, Skeleton, Stack, Typography } from '@mui/material'
 import { motion } from 'framer-motion'
-import CheckCircleOutlineRoundedIcon from '@mui/icons-material/CheckCircleOutlineRounded'
 import CheckCircleRoundedIcon from '@mui/icons-material/CheckCircleRounded'
-import ArrowForwardRoundedIcon from '@mui/icons-material/ArrowForwardRounded'
-import ErrorOutlineRoundedIcon from '@mui/icons-material/ErrorOutlineRounded'
-import HelpOutlineRoundedIcon from '@mui/icons-material/HelpOutlineRounded'
-import KeyboardArrowDownRoundedIcon from '@mui/icons-material/KeyboardArrowDownRounded'
-import RemoveCircleOutlineRoundedIcon from '@mui/icons-material/RemoveCircleOutlineRounded'
 import UnfoldLessRoundedIcon from '@mui/icons-material/UnfoldLessRounded'
 import UnfoldMoreRoundedIcon from '@mui/icons-material/UnfoldMoreRounded'
-import WarningAmberRoundedIcon from '@mui/icons-material/WarningAmberRounded'
-import Link from 'next/link'
 import type { EvidenceItem, ScanContext, ScanResult, SecurityGrade, StrengthLevel } from '@/features/security/types'
-import { SecurityFeature, type SecurityContract } from '@/features/security'
+import { SecurityFeature } from '@/features/security'
 import { useLoadFeature } from '@/features/__core__'
 import { shortenAddress } from '@safe-global/utils/utils/formatters'
+import {
+  EvidenceList,
+  Row,
+  StatusIcon,
+  buildExpanded,
+  isPassingStatus,
+  makeBuildCta,
+  sortBySeverity,
+  type SectionRow,
+} from './primitives'
+import { useSecurityChecks } from './useSecurityChecks'
 
 type SecurityPanelViewProps = {
   scanContext: ScanContext | null
@@ -37,220 +40,6 @@ const GRADE_BG_BY_STRENGTH: Record<StrengthLevel, string> = {
   Moderate: 'warning.background',
   Weak: 'error.background',
   Critical: 'error.background',
-}
-
-/**
- * A row is considered "passing" (bucketed into the accordion) when the user has no action to take.
- * `inconclusive` means "we couldn't determine" (e.g. 3rd-party screening API unavailable) — treating
- * it as passing avoids false alarm; the row is still expandable with its distinct grey icon.
- */
-const isPassingStatus = (s: ScanResult['status']) => s === 'clear' || s === 'not_applicable' || s === 'inconclusive'
-
-const SEVERITY_RANK: Record<SecurityGrade, number> = { Critical: 0, High: 1, Medium: 2, Low: 3 }
-
-/** Sort row entries with most severe first, falling back to original order. */
-const sortBySeverity = <T extends { severity: SecurityGrade }>(items: T[]): T[] =>
-  [...items].sort((a, b) => SEVERITY_RANK[a.severity] - SEVERITY_RANK[b.severity])
-
-// ───────────────────────────────────────────────────────────────────────────────
-// Unified row + chip + section primitives
-
-/** Leading icon reflecting a check's status. Sized & colored to match the accordion summary icon. */
-const StatusIcon = ({ status }: { status: ScanResult['status'] }): ReactElement => {
-  const sx = { fontSize: 18, flexShrink: 0 } as const
-  if (status === 'clear') return <CheckCircleOutlineRoundedIcon sx={{ ...sx, color: 'success.main' }} />
-  if (status === 'not_applicable') return <RemoveCircleOutlineRoundedIcon sx={{ ...sx, color: 'text.secondary' }} />
-  if (status === 'inconclusive') return <HelpOutlineRoundedIcon sx={{ ...sx, color: 'text.disabled' }} />
-  if (status === 'partial') return <WarningAmberRoundedIcon sx={{ ...sx, color: 'warning.main' }} />
-  // issue
-  return <ErrorOutlineRoundedIcon sx={{ ...sx, color: 'error.main' }} />
-}
-
-type RowProps = {
-  /** Leading status icon (same visual weight as accordion summary icon) */
-  leadIcon?: ReactNode
-  title: string
-  /** Trailing action node (used by modules summary's "View N" button) */
-  trailing?: ReactNode
-  /** Reveal additional content on click */
-  expandedContent?: ReactNode
-}
-
-const Row = ({ leadIcon, title, trailing, expandedContent }: RowProps): ReactElement => {
-  const [expanded, setExpanded] = useState(false)
-  const expandable = !!expandedContent
-
-  return (
-    <Box
-      sx={{
-        px: 2,
-        cursor: expandable ? 'pointer' : 'default',
-        '&:hover': expandable ? { backgroundColor: 'action.hover' } : {},
-      }}
-      onClick={expandable ? () => setExpanded((v) => !v) : undefined}
-    >
-      <Stack direction="row" spacing={1.25} alignItems="center" sx={{ py: 1.25 }}>
-        {leadIcon && <Box sx={{ display: 'flex', alignItems: 'center', flexShrink: 0 }}>{leadIcon}</Box>}
-        <Typography
-          variant="body2"
-          fontWeight={600}
-          sx={{ flex: 1, minWidth: 0, overflow: 'hidden', textOverflow: 'ellipsis' }}
-          noWrap
-          title={title}
-        >
-          {title}
-        </Typography>
-        {trailing}
-        {expandable && (
-          <KeyboardArrowDownRoundedIcon
-            sx={{
-              color: 'text.secondary',
-              fontSize: 18,
-              flexShrink: 0,
-              transform: expanded ? 'rotate(180deg)' : 'rotate(0deg)',
-              transition: 'transform 0.2s',
-            }}
-          />
-        )}
-      </Stack>
-      {expandable && (
-        <Collapse in={expanded}>
-          {/* Align expanded body's left edge with the title (icon 18px + gap 1.25 = 3.5 spacing units). */}
-          <Box sx={{ pb: 1.5, pl: leadIcon ? 3.5 : 0 }} onClick={(e) => e.stopPropagation()}>
-            {typeof expandedContent === 'string' ? (
-              <Typography variant="caption" color="text.secondary" sx={{ lineHeight: 1.5, display: 'block' }}>
-                {expandedContent}
-              </Typography>
-            ) : (
-              expandedContent
-            )}
-          </Box>
-        </Collapse>
-      )}
-    </Box>
-  )
-}
-
-type Cta = { label: string; href: string }
-
-/**
- * Factory that binds `checkDefs` to a CTA builder. Consumers obtain `checkDefs`
- * via useLoadFeature and pass it once; the returned function is then used at
- * each render site.
- *
- * The returned CTA builder returns null when:
- *  - the row is passing (no action needed),
- *  - we don't yet have a `safeQueryParam` (chain metadata still loading), or
- *  - there's no checkDefs entry for this check id.
- * Label precedence: `ScanResult.ctaLabelOverride` → `checkDefs[id].ctaLabel`.
- */
-const makeBuildCta =
-  (checkDefs: SecurityContract['checkDefs']) =>
-  (checkId: string, result: ScanResult | undefined, safeQueryParam: string | undefined): Cta | null => {
-    if (!safeQueryParam) return null
-    const def = checkDefs[checkId]
-    if (!def) return null
-    if (result && isPassingStatus(result.status)) return null
-    const label = result?.ctaLabelOverride || def.ctaLabel
-    return { label, href: `${def.fixRoute}?safe=${encodeURIComponent(safeQueryParam)}` }
-  }
-
-const CtaLink = ({ cta }: { cta: Cta }): ReactElement => (
-  <Link
-    href={cta.href}
-    onClick={(e) => e.stopPropagation()}
-    style={{ textDecoration: 'none', alignSelf: 'flex-start', borderRadius: 4 }}
-    className="security-cta-link"
-  >
-    <Stack
-      direction="row"
-      spacing={0.75}
-      alignItems="center"
-      sx={{
-        color: 'primary.main',
-        cursor: 'pointer',
-        transition: 'color 0.15s',
-        '& .cta-icon': { transition: 'transform 0.15s' },
-        '&:hover': { color: 'primary.dark' },
-        '&:hover .cta-label': { textDecoration: 'underline' },
-        '&:hover .cta-icon': { transform: 'translateX(2px)' },
-        '.security-cta-link:focus-visible &': {
-          outline: '2px solid',
-          outlineColor: 'primary.main',
-          outlineOffset: 2,
-          borderRadius: 4,
-        },
-      }}
-    >
-      <Typography className="cta-label" variant="caption" fontWeight={700} sx={{ color: 'inherit', lineHeight: 1.5 }}>
-        {cta.label}
-      </Typography>
-      <ArrowForwardRoundedIcon className="cta-icon" sx={{ fontSize: 14, color: 'inherit' }} />
-    </Stack>
-  </Link>
-)
-
-/** Expanded-row body: optional intro paragraph + evidence key/value list + optional CTA. */
-const EvidenceList = ({
-  intro,
-  evidence,
-  cta,
-}: {
-  intro?: string
-  evidence?: EvidenceItem[]
-  cta?: Cta | null
-}): ReactElement | null => {
-  const hasEvidence = !!evidence && evidence.length > 0
-  if (!intro && !hasEvidence && !cta) return null
-  return (
-    <Stack spacing={0.75}>
-      {intro && (
-        <Typography variant="caption" color="text.primary" sx={{ lineHeight: 1.5, display: 'block' }}>
-          {intro}
-        </Typography>
-      )}
-      {hasEvidence && (
-        <Stack spacing={0.25}>
-          {evidence!.map((item, idx) =>
-            typeof item === 'string' ? (
-              <Typography
-                key={idx}
-                variant="caption"
-                color="text.secondary"
-                sx={{ lineHeight: 1.5, display: 'block', wordBreak: 'break-word' }}
-              >
-                {item}
-              </Typography>
-            ) : (
-              <Stack key={idx} direction="row" spacing={1} alignItems="baseline">
-                <Typography
-                  variant="caption"
-                  color="text.secondary"
-                  sx={{ minWidth: 96, flexShrink: 0, lineHeight: 1.5 }}
-                >
-                  {item.label}
-                </Typography>
-                <Typography variant="caption" color="text.primary" sx={{ lineHeight: 1.5, wordBreak: 'break-word' }}>
-                  {item.value}
-                </Typography>
-              </Stack>
-            ),
-          )}
-        </Stack>
-      )}
-      {cta && <CtaLink cta={cta} />}
-    </Stack>
-  )
-}
-
-/** Build expanded body for a row backed by a ScanResult. Returns undefined when there's nothing to show. */
-const buildExpanded = (result: ScanResult | undefined, cta?: Cta | null): ReactNode => {
-  if (!result) return cta ? <EvidenceList cta={cta} /> : undefined
-  const isOk = result.status === 'clear' || result.status === 'not_applicable'
-  const intro = !isOk && result.remediation ? result.remediation : undefined
-  const hasEvidence = result.evidence && result.evidence.length > 0
-  if (!intro && !hasEvidence && !cta) return undefined
-  return <EvidenceList intro={intro} evidence={result.evidence} cta={cta} />
 }
 
 const MotionBox = motion.create(Box)
@@ -417,8 +206,6 @@ const SignersSection = ({
 
   if (!security.$isReady || !buildCta) return null
 
-  type SectionRow = { key: string; severity: SecurityGrade; isPassing: boolean; node: ReactNode }
-
   // Bucket per-signer rows into flagged vs screened.
   const signerItems: SectionRow[] = []
   // All signer rows route back to the signer_integrity remediation page.
@@ -548,290 +335,11 @@ const SecurityChecksSection = ({
   results: Record<string, ScanResult>
   safeQueryParam?: string
 }): ReactElement | null => {
-  const security = useLoadFeature(SecurityFeature)
-  const buildCta = useMemo(
-    () => (security.$isReady ? makeBuildCta(security.checkDefs) : null),
-    [security.$isReady, security.checkDefs],
-  )
-
-  const accountSetupResult = results['account_setup']
-  const recoveryResult = results['recovery']
-  const versionResult = results['contract_version']
-  const factoryResult = results['factory_validation']
-  const guardResult = results['guard']
-  const fallbackResult = results['fallback_handler']
-  const modulesResult = results['modules']
-  const pendingResult = results['pending_tx']
-  const scanningResult = results['transaction_scanning']
-
-  // Modules collapse: when N > 2, summarize and let user expand
-  const [modulesExpanded, setModulesExpanded] = useState(false)
-  // Passing checks accordion
+  const { isReady, failingRows, passingRows } = useSecurityChecks(scanContext, results, safeQueryParam)
   const [passingExpanded, setPassingExpanded] = useState(false)
 
   // Feature not yet loaded — render nothing; the panel skeleton covers this state.
-  // This guard runs AFTER all hooks to respect Rules of Hooks.
-  if (!security.$isReady || !buildCta) return null
-
-  const ZERO_ADDRESS = security.zeroAddress
-  const hasGuard = scanContext.guard !== null && scanContext.guard.value !== ZERO_ADDRESS
-  const hasFallback = scanContext.fallbackHandler !== null && scanContext.fallbackHandler.value !== ZERO_ADDRESS
-  const activeModules = (scanContext.modules ?? []).filter((m) => m.value !== ZERO_ADDRESS)
-
-  const showModuleSummary = activeModules.length > 2 && !modulesExpanded
-
-  type SectionRow = { key: string; severity: SecurityGrade; isPassing: boolean; node: ReactNode }
-  const items: SectionRow[] = []
-
-  const iconFor = (r: ScanResult) => <StatusIcon status={r.status} />
-
-  if (accountSetupResult) {
-    const ok = isPassingStatus(accountSetupResult.status)
-    const title = ok
-      ? 'Signing threshold is strong'
-      : scanContext.threshold === 1
-        ? 'Single signer controls this Safe'
-        : 'Signing threshold is low'
-    items.push({
-      key: 'threshold',
-      severity: accountSetupResult.severity,
-      isPassing: ok,
-      node: (
-        <Row
-          leadIcon={iconFor(accountSetupResult)}
-          title={title}
-          expandedContent={buildExpanded(
-            accountSetupResult,
-            buildCta('account_setup', accountSetupResult, safeQueryParam),
-          )}
-        />
-      ),
-    })
-  }
-
-  if (recoveryResult) {
-    const ok = isPassingStatus(recoveryResult.status)
-    const title =
-      recoveryResult.status === 'clear'
-        ? 'Recovery is configured'
-        : recoveryResult.status === 'not_applicable'
-          ? 'Recovery not available on this network'
-          : 'Recovery is not configured'
-    items.push({
-      key: 'recovery',
-      severity: recoveryResult.severity,
-      isPassing: ok,
-      node: (
-        <Row
-          leadIcon={iconFor(recoveryResult)}
-          title={title}
-          expandedContent={buildExpanded(recoveryResult, buildCta('recovery', recoveryResult, safeQueryParam))}
-        />
-      ),
-    })
-  }
-
-  if (versionResult) {
-    const ok = isPassingStatus(versionResult.status)
-    const title = ok ? 'Contract version is up to date' : 'Contract version is outdated'
-    items.push({
-      key: 'version',
-      severity: versionResult.severity,
-      isPassing: ok,
-      node: (
-        <Row
-          leadIcon={iconFor(versionResult)}
-          title={title}
-          expandedContent={buildExpanded(versionResult, buildCta('contract_version', versionResult, safeQueryParam))}
-        />
-      ),
-    })
-  }
-
-  if (factoryResult) {
-    const ok = isPassingStatus(factoryResult.status)
-    const title = ok ? 'Deployed via official Safe factory' : 'Deployed from an unrecognized source'
-    items.push({
-      key: 'factory',
-      severity: factoryResult.severity,
-      isPassing: ok,
-      node: (
-        <Row
-          leadIcon={iconFor(factoryResult)}
-          title={title}
-          expandedContent={buildExpanded(factoryResult, buildCta('factory_validation', factoryResult, safeQueryParam))}
-        />
-      ),
-    })
-  }
-
-  if (guardResult) {
-    const ok = isPassingStatus(guardResult.status)
-    const title = ok
-      ? hasGuard
-        ? 'Transaction guard is active'
-        : 'No transaction guard in use'
-      : hasGuard
-        ? 'Transaction guard is unverified'
-        : 'Transaction guard is recommended'
-    items.push({
-      key: 'guard',
-      severity: guardResult.severity,
-      isPassing: ok,
-      node: (
-        <Row
-          leadIcon={iconFor(guardResult)}
-          title={title}
-          expandedContent={buildExpanded(guardResult, buildCta('guard', guardResult, safeQueryParam))}
-        />
-      ),
-    })
-  }
-
-  if (fallbackResult) {
-    const ok = isPassingStatus(fallbackResult.status)
-    // Scanner emits a human label like "Official Safe fallback handler" / "CoW Protocol TWAP handler"
-    // in evidence — reuse it directly so the title auto-matches each variant.
-    const handlerLabel = fallbackResult.evidence?.find(
-      (e): e is { label: string; value: string } => typeof e !== 'string' && e.label === 'Status',
-    )?.value
-    const title = ok
-      ? hasFallback
-        ? handlerLabel || 'Fallback handler is active'
-        : 'No fallback handler in use'
-      : 'Fallback handler is unverified'
-    items.push({
-      key: 'fallback',
-      severity: fallbackResult.severity,
-      isPassing: ok,
-      node: (
-        <Row
-          leadIcon={iconFor(fallbackResult)}
-          title={title}
-          expandedContent={buildExpanded(fallbackResult, buildCta('fallback_handler', fallbackResult, safeQueryParam))}
-        />
-      ),
-    })
-  }
-
-  if (modulesResult) {
-    if (activeModules.length === 0) {
-      items.push({
-        key: 'modules-empty',
-        severity: modulesResult.severity,
-        isPassing: isPassingStatus(modulesResult.status),
-        node: (
-          <Row
-            leadIcon={iconFor(modulesResult)}
-            title="No modules installed"
-            expandedContent={buildExpanded(modulesResult, buildCta('modules', modulesResult, safeQueryParam))}
-          />
-        ),
-      })
-    } else if (showModuleSummary) {
-      // Collapsed summary row — not expandable, acts as a gateway to per-module rows.
-      items.push({
-        key: 'modules-summary',
-        severity: modulesResult.severity,
-        isPassing: isPassingStatus(modulesResult.status),
-        node: (
-          <Row
-            leadIcon={iconFor(modulesResult)}
-            title={`Modules & Extensions · ${activeModules.length} installed`}
-            trailing={
-              <Button
-                size="small"
-                variant="text"
-                onClick={(e) => {
-                  e.stopPropagation()
-                  setModulesExpanded(true)
-                }}
-                sx={{ fontSize: '0.7rem', p: 0, minWidth: 0, textTransform: 'none', fontWeight: 600 }}
-              >
-                View all
-              </Button>
-            }
-          />
-        ),
-      })
-    } else {
-      const modulesCta = buildCta('modules', modulesResult, safeQueryParam)
-      activeModules.forEach((mod) => {
-        const trusted = security.isKnownModuleByName(mod.name)
-        const severity: SecurityGrade = trusted ? 'Low' : 'High'
-        const status: ScanResult['status'] = trusted ? 'clear' : 'issue'
-        // Show the module's name as the title for both trusted and unrecognized modules — users
-        // need to identify *which* module when it's flagged. The icon + intro convey the verdict.
-        const title = mod.name || shortenAddress(mod.value)
-        const perModuleEvidence: EvidenceItem[] = [
-          { label: 'Address', value: mod.value },
-          ...(mod.name ? [{ label: 'Name', value: mod.name }] : []),
-        ]
-        const intro = trusted
-          ? 'Recognized Safe ecosystem module.'
-          : "Unrecognized module — not in the known Safe ecosystem deployments. Review carefully and remove if you don't recognize it."
-        items.push({
-          key: `module-${mod.value}`,
-          severity,
-          isPassing: trusted,
-          node: (
-            <Row
-              leadIcon={<StatusIcon status={status} />}
-              title={title}
-              expandedContent={
-                <EvidenceList intro={intro} evidence={perModuleEvidence} cta={trusted ? null : modulesCta} />
-              }
-            />
-          ),
-        })
-      })
-    }
-  }
-
-  if (scanningResult) {
-    const ok = isPassingStatus(scanningResult.status)
-    const title = ok ? 'Transaction scanning is enabled' : 'Transaction scanning is disabled'
-    items.push({
-      key: 'scanning',
-      severity: scanningResult.severity,
-      isPassing: ok,
-      node: (
-        <Row
-          leadIcon={iconFor(scanningResult)}
-          title={title}
-          expandedContent={buildExpanded(
-            scanningResult,
-            buildCta('transaction_scanning', scanningResult, safeQueryParam),
-          )}
-        />
-      ),
-    })
-  }
-
-  if (pendingResult) {
-    const ok = isPassingStatus(pendingResult.status)
-    const queued = scanContext.queuedTxCount
-    const title = ok
-      ? queued > 0
-        ? 'Queue is up to date'
-        : 'No pending transactions'
-      : 'Pending transactions are stale'
-    items.push({
-      key: 'pending',
-      severity: pendingResult.severity,
-      isPassing: ok,
-      node: (
-        <Row
-          leadIcon={iconFor(pendingResult)}
-          title={title}
-          expandedContent={buildExpanded(pendingResult, buildCta('pending_tx', pendingResult, safeQueryParam))}
-        />
-      ),
-    })
-  }
-
-  const failingRows = sortBySeverity(items.filter((i) => !i.isPassing)).map(({ key, node }) => ({ key, node }))
-  const passingRows = items.filter((i) => i.isPassing).map(({ key, node }) => ({ key, node }))
+  if (!isReady) return null
 
   const footer =
     passingRows.length > 0 ? (
